Avoid ORD violations in *Message() functions
